requirements.txt
Flask
mysql-connector-python
python-dotenv
flasgger
apscheduler
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.7
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.54.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
gdown==5.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
inflate64==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
labelme==5.8.1
lark==1.1.9
lazy_loader==0.4
lmdb==1.6.2
loguru==0.7.3
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multivolumefile==0.2.3
mysql-connector-python==9.2.0
natsort==8.4.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
oletools==0.60.2
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
osam==0.2.3
packaging==24.2
paddleocr==2.10.0
paddlepaddle-gpu==2.6.2
pandas==2.2.3
pcodedmp==1.2.6
pdf2image==1.17.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
py7zr==0.22.0
pybcj==1.0.3
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pyppmd==1.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.16.2
qrdet==2.5
quadrilateral-fitter==1.12
RapidFuzz==3.12.2
rarfile==4.2
red-black-tree-mod==1.22
regex==2024.11.6
requests==2.32.3
RTFDE==0.1.2
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
supervisor==4.2.5
sympy==1.13.1
termcolor==3.0.1
texttable==1.7.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.103
ultralytics-thop==2.0.14
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3